requirements.txt
pyyaml
bs4
openai
async_timeout
python-telegram-bot==20.7
tiktoken
aiohttp
pdfminer.six